refactor(#10) - id int -> Long type 반환

+++ src/main/java/com/flab/infrun/lecture/infrastructure/persistence/mybatis/mapper/LectureDetailMyBatisMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface LectureDetailMyBatisMapper {
 
-    int save(LectureDetail entity);
+    long save(LectureDetail entity);
 
     Optional<LectureDetail> findById(Long id);
 }
