+++ src/test/java/com/flab/infrun/lecture/infrastructure/persistence/LectureVideoFileRepositoryTest.java
@@ -21,9 +21,9 @@ class LectureVideoFileRepositoryTest {
     void save() {
         LectureVideoFile lectureVideoFile = LectureVideoFile.of("/lecture", "파일명1.pdf");
 
-        int saved = mapper.save(lectureVideoFile);
+        long saved = mapper.save(lectureVideoFile);
 
-        Optional<LectureVideoFile> lecture1 = mapper.findById((long) saved);
+        Optional<LectureVideoFile> lecture1 = mapper.findById(saved);
         assertThat(lecture1.get()).isEqualTo(lectureVideoFile);
     }
 }
